Make IOChannel implementations reflect the documentation more. Drop error codes for a boolean bad(). Use streamsize for bytes written to / read from streams, streampos for positions, and size_t for stream size, which is logically correct, consistent, and reduces unnecessary casting.
Silence plugin debugging again.

consumeInput() should also apparently update bytesTotal.

+++ libcore/parser/SWFMovieDefinition.h
@@ -511,7 +511,7 @@ private:
 	std::auto_ptr<IOChannel> _in;
 
 	/// swf end position (as read from header)
-	unsigned int _swf_end_pos;
+    size_t _swf_end_pos;
 
 	/// asyncronous SWF loader and parser
 	MovieLoader _loader;
